# openapi.yml
openapi: 3.0.3
info:
  title: EasyFinder API
  version: 1.1.0
  description: >
    EasyFinder API contract aligned to currently implemented routes in apps/api.
    Updated to include IronPlanet scraping ingestion and correct NDA/Auth behaviors.
servers:
  - url: https://easyfinder.fly.dev
    description: Production
  - url: http://localhost:8080
    description: Local
...
paths:
...
  /api/demo/listings/{id}:
    get:
      tags: [Demo Listings]
      summary: Get demo listing by id
      parameters:
        - $ref: "#/components/parameters/ListingId"
      responses:
        "200":
          description: Demo listing detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingResponse"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /api/listings:
    get:
      tags: [Listings]
      summary: Get ranked live listings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Live listing array
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingsResponse"
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "402": { $ref: "#/components/responses/PaymentRequiredError" }
        "403": { $ref: "#/components/responses/NdaRequiredError" }
        "503": { $ref: "#/components/responses/DemoModeActiveError" }
    post:
      tags: [Listings]
      summary: Create listing (not implemented)
      security:
        - bearerAuth: []
      responses:
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "403":
          description: NDA required or writes disabled in demo mode
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "501":
          description: Not implemented
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    put:
      tags: [Listings]
      summary: Update listing (not implemented)
      security:
        - bearerAuth: []
      responses:
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "403":
          description: NDA required or writes disabled in demo mode
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "501":
          description: Not implemented
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    delete:
      tags: [Listings]
      summary: Delete listing (not implemented)
      security:
        - bearerAuth: []
      responses:
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "403":
          description: NDA required or writes disabled in demo mode
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "501":
          description: Not implemented
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /api/listings/{id}:
    get:
      tags: [Listings]
      summary: Get live listing by id
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ListingId"
      responses:
        "200":
          description: Live listing detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingResponse"
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "402": { $ref: "#/components/responses/PaymentRequiredError" }
        "403": { $ref: "#/components/responses/NdaRequiredError" }
        "404": { $ref: "#/components/responses/NotFoundError" }
        "503": { $ref: "#/components/responses/DemoModeActiveError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ListingId:
      in: path
      name: id
      required: true
      schema: { type: string }
...
  responses:
    UnauthorizedError:
      description: Authentication required.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    NotFoundError:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    PaymentRequiredError:
      description: Active subscription required.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    NdaRequiredError:
      description: NDA must be accepted.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    DemoModeActiveError:
      description: Live listings unavailable in demo mode.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
        requestId: { type: string }
...
    Listing:
      type: object
      required: [id, title, category, state, source, operable]
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        category: { type: string }
        state: { type: string }
        price: { type: number, nullable: true }
        hours: { type: number, nullable: true }
        year: { type: integer, nullable: true }
        condition: { type: number, nullable: true }
        sellerType: { type: string, enum: [dealer, auction, private, unknown], nullable: true }
        shippingAvailable: { type: boolean, nullable: true }
        availability: { type: string, enum: [in_stock, scheduled_auction, unknown], nullable: true }
        lastSeenAt: { type: string, format: date-time, nullable: true }
        listingUpdatedAt: { type: string, format: date-time, nullable: true }
        photoCount: { type: integer, nullable: true }
        hasInspectionReport: { type: boolean, nullable: true }
        hasServiceHistory: { type: boolean, nullable: true }
        verifiedSeller: { type: boolean, nullable: true }
        city: { type: string, nullable: true }
        country: { type: string, nullable: true }
        operable: { type: boolean }
        source: { type: string }
        images:
          type: array
          items: { type: string }
        imageUrl: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        totalScore: { type: number }
        scoreV2:
          $ref: "#/components/schemas/ScorePillars"
        confidenceScore: { type: number }
        reasons:
          type: array
          items: { $ref: "#/components/schemas/ScoreReason" }
        flags:
          type: array
          items: { type: string }
        bestOptionEligible: { type: boolean }
        score:
          $ref: "#/components/schemas/ScoreBreakdown"
    ScorePillars:
      type: object
      properties:
        deal: { type: number }
        usage: { type: number }
        risk: { type: number }
        speed: { type: number }
        quality: { type: number }
    ScoreReason:
      type: object
      properties:
        kind: { type: string, enum: [deal, usage, risk, speed, quality] }
        message: { type: string }
    ScoreBreakdown:
      type: object
      properties:
        total: { type: number }
        breakdown:
          $ref: "#/components/schemas/ScorePillars"
        scoreV2:
          $ref: "#/components/schemas/ScorePillars"
        reasons:
          type: array
          items: { $ref: "#/components/schemas/ScoreReason" }
        flags:
          type: array
          items: { type: string }
        confidence: { type: number }
        confidenceScore: { type: number }
        bestOptionEligible: { type: boolean }
        disqualified: { type: boolean }
    ListingsResponse:
      type: object
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/Listing" }
        requestId: { type: string }
    ListingResponse:
      type: object
      properties:
        data: { $ref: "#/components/schemas/Listing" }
        requestId: { type: string }
...